--- a/compiler/runtime.go
+++ b/compiler/runtime.go
@@ -696,7 +696,7 @@ func (f *functionValue) Call(ctx context.Context, args values.Object) (values.Va
 		scope.Set(p.Key, a)
 	}
 
-	fn, err := Compile(scope, f.fn, args.Type())
+	fn, err := Compile(ctx, scope, f.fn, args.Type())
 	if err != nil {
 		return nil, err
 	}