--- a/internal/compiler/function.go
+++ b/internal/compiler/function.go
@@ -1,0 +1,1 @@
+package compiler