--- a/internal/compiler/value_mapper.go
+++ b/internal/compiler/value_mapper.go
@@ -1,0 +1,11 @@
+package compiler
+
+type valueMapper interface {
+	Map(in Value, scope []Value)
+}
+
+type basicValueMapper int
+
+func (b basicValueMapper) Map(in Value, scope []Value) {
+	scope[b] = in
+}