… compiler/compiler_test.go
… execute/row_fn_test.go
… internal/compiler/binary.gen.go
… internal/compiler/binary.go
… internal/compiler/compiler_test.go
… stdlib/experimental/join.go
… stdlib/generate/from.go
… stdlib/influxdata/influxdb/to.go
… stdlib/universe/filter.go
… stdlib/universe/map.go
… stdlib/universe/reduce.go
… stdlib/universe/schema_mutators.go
… stdlib/universe/state_tracking.go
… stdlib/universe/table_fns.go